Model Parameters for Cumba

This dataset contains the default model parameters used in the Cumba package. Each element corresponds to a model parameter and includes its description, current value, minimum, and maximum.

Usage

cumbaParameters

Format

A named list where each entry is a list with:

description

Text description of the parameter

value

The default value of the parameter

min

The minimum allowed value

max

The maximum allowed value

Examples

names(cumbaParameters)
#>  [1] "Tbase"                        "Topt"                        
#>  [3] "Tmax"                         "Theat"                       
#>  [5] "Tcold"                        "FIntMax"                     
#>  [7] "CycleLength"                  "TransplantingLag"            
#>  [9] "FloweringLag"                 "HalfIntGrowth"               
#> [11] "HalfIntSenescence"            "InitialInt"                  
#> [13] "RUE"                          "KcIni"                       
#> [15] "KcMax"                        "RootIncrease"                
#> [17] "RootDepthMax"                 "RootDepthInitial"            
#> [19] "SoilWaterInitial"             "FieldCapacity"               
#> [21] "WiltingPoint"                 "DepletionFraction"           
#> [23] "WaterStressSensitivity"       "FloweringSlope"              
#> [25] "FloweringMax"                 "k0"                          
#> [27] "FruitWaterContentMin"         "FruitWaterContentMax"        
#> [29] "FruitWaterContentInc"         "FruitWaterContentDecreaseMax"
cumbaParameters$Tbase$value
#> [1] 10