Tomato field trial data from Foggia (Italy)

A dataset containing weather, irrigation, management, and production data from processing tomato field trials conducted near Foggia (Apulia, Italy). The dataset is organized as a list of four data frames.

Usage

tomatoFoggia

Format

A list of four data frames:

weather

Daily weather data with columns: DATE (POSIXct), Tx (max temp, °C), Tn (min temp, °C), P (precipitation, mm), Rad (radiation, MJ m-2 d-1), RHx (max RH, Date (Date), Site ("Foggia"), Lat (latitude, numeric).

irrigation

Irrigation records with columns: ID (plot id), DATE (POSIXct), WVOL (mm), YEAR, CV (cultivar), TRANS (transplanting, POSIXct), HARV (harvest, POSIXct), IRR_M (management, e.g. "50 Site ("Foggia").

management

Plot identifiers and treatments with columns: ID, YEAR, CV, TRANS (POSIXct), HARV (POSIXct), IRR_M.

production

Production and quality data with columns: ID, yield_ref (t ha-1), yield_ref_sd (t ha-1), brix_ref (°Brix), brix_ref_sd (°Brix).

Source

Carucci et al. (2024) https://doi.org/10.1016/j.dib.2024.110225

Examples

data(tomatoFoggia)
str(tomatoFoggia)
#> List of 4
#>  $ weather   : tibble [949 × 11] (S3: tbl_df/tbl/data.frame)
#>   ..$ DATE: POSIXct[1:949], format: "2005-04-28" "2005-04-29" ...
#>   ..$ Tx  : num [1:949] 27.7 23.6 23.2 27.6 29.4 32 30.1 24.8 17.2 23.5 ...
#>   ..$ Tn  : num [1:949] 7.2 10.4 8.8 7.4 8.8 10.8 12.2 13.5 12.1 10.4 ...
#>   ..$ P   : num [1:949] 0 0 0 0 0 0 0 0 1.2 0 ...
#>   ..$ Rad : num [1:949] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ RHx : num [1:949] 91 91.5 78.5 88 92 89.5 79.5 88 94.5 86.5 ...
#>   ..$ RHn : num [1:949] 32.5 35.5 33.5 25 32 30.5 24.5 40 74.5 29 ...
#>   ..$ W   : num [1:949] 0.0472 0.1056 2.5 1.6 1.3 ...
#>   ..$ Date: Date[1:949], format: "2005-04-28" "2005-04-29" ...
#>   ..$ Site: chr [1:949] "Foggia" "Foggia" "Foggia" "Foggia" ...
#>   ..$ Lat : num [1:949] 41 41 41 41 41 41 41 41 41 41 ...
#>  $ irrigation: tibble [524 × 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ ID   : num [1:524] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ DATE : POSIXct[1:524], format: "2005-04-28" "2005-05-05" ...
#>   ..$ WVOL : num [1:524] 27.9 20 10.5 10.5 15.8 ...
#>   ..$ YEAR : num [1:524] 2005 2005 2005 2005 2005 ...
#>   ..$ CV   : chr [1:524] "Ulisse" "Ulisse" "Ulisse" "Ulisse" ...
#>   ..$ TRANS: POSIXct[1:524], format: "2005-04-28" "2005-04-28" ...
#>   ..$ HARV : POSIXct[1:524], format: "2005-07-31" "2005-07-31" ...
#>   ..$ IRR_M: chr [1:524] "50%" "50%" "50%" "50%" ...
#>   ..$ Site : chr [1:524] "Foggia" "Foggia" "Foggia" "Foggia" ...
#>  $ management: tibble [32 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ ID   : num [1:32] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ YEAR : num [1:32] 2005 2005 2005 2005 2005 ...
#>   ..$ CV   : chr [1:32] "Ulisse" "Ulisse" "Ulisse" "Ulisse" ...
#>   ..$ TRANS: POSIXct[1:32], format: "2005-04-28" "2005-04-28" ...
#>   ..$ HARV : POSIXct[1:32], format: "2005-07-31" "2005-07-31" ...
#>   ..$ IRR_M: chr [1:32] "50%" "75%" "100%" "507550%" ...
#>  $ production: tibble [32 × 5] (S3: tbl_df/tbl/data.frame)
#>   ..$ ID          : num [1:32] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ yield_ref   : num [1:32] 69.1 87.3 87.6 81 77.4 ...
#>   ..$ yield_ref_sd: num [1:32] 4.04 11.1 3.68 4.9 4.55 ...
#>   ..$ brix_ref    : num [1:32] 5.8 5.2 5 5 5.17 ...
#>   ..$ brix_ref_sd : num [1:32] 0.361 0.3 0.173 0.2 0.551 ...
head(tomatoFoggia$weather)
#> # A tibble: 6 × 11
#>   DATE                   Tx    Tn     P   Rad   RHx   RHn      W Date      
#>   <dttm>              <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <date>    
#> 1 2005-04-28 00:00:00  27.7   7.2     0     0  91    32.5 0.0472 2005-04-28
#> 2 2005-04-29 00:00:00  23.6  10.4     0     0  91.5  35.5 0.106  2005-04-29
#> 3 2005-04-30 00:00:00  23.2   8.8     0     0  78.5  33.5 2.5    2005-04-30
#> 4 2005-05-01 00:00:00  27.6   7.4     0     0  88    25   1.6    2005-05-01
#> 5 2005-05-02 00:00:00  29.4   8.8     0     0  92    32   1.3    2005-05-02
#> 6 2005-05-03 00:00:00  32    10.8     0     0  89.5  30.5 0.0632 2005-05-03
#> # ℹ 2 more variables: Site <chr>, Lat <dbl>