the CUMBA model

It is a daily time step simulation model which computes the yield and brix degree of a tomato crop as a function of weather data and irrigation options.

Usage

cumba_scenario(
  weather,
  param,
  estimateRad = T,
  estimateET0 = T,
  transplantingDOY = 120,
  waterStressLevel = 0.5,
  minimumTurn = 4,
  fullOut = F
)

Arguments

weather

a dataframe with weather data which must have the following columns......

param

A named list of model parameters, where each element is a list with fields 'value', 'min', 'max', and 'description'

estimateRad

a boolean value to estimate solar radiation based on temperature using Hargreaves model. Default to 'true' (implying that the column Lat is present in weather df) if 'false' the 'weather' df must have the Rad column

estimateET0

a boolean value to estimate reference evapotranspiration based on temperature using Hargreaves model. Default to 'true'

waterStressLevel

a float corresponding to the threshold of water stress to trigger automatic irrigation. Default to .5, it is needed only if deficitIrrigation is 'true'.

minimumTurn

an integer corresponding to the minimum number of days elapsed from the previous irrigation event. Default to 4, it is needed only if deficitIrrigation is 'true'.

fullOut

boolean, if FALSE main output variables are saved (default), if TRUE all variables are saved

deficitIrrigation

a boolean value to estimate irrigation requirements. Default to 'false', implying that the irrigation_df is provided.

Value

a dataframe containing the weatherDf plus the daily outputs of the cumba model

Examples

#' # Example weather dataframe
weather <- data.frame(
  Site = "TestSite",
  Tx = c(30, 32, 31),
  Tn = c(20, 21, 19),
  P = c(0, 5, 2),
  DATE = as.Date(c("2025-06-01", "2025-06-02", "2025-06-03")),
  Lat = 40
)

# Minimal parameters: use default package data
params <- lapply(cumbaParameters, function(p) p)  # copy default parameters

# Run the model in automatic irrigation (scenario) mode
# result <- cumba_scenario(
#   weather,
#   params,
#   estimateRad = TRUE,
#   estimateET0 = TRUE,
#   waterStressLevel = 0.5,
#   minimumTurn = 4
# )